+++ Virtuoso/CHANGELOG.md
@@ -2,10 +2,15 @@
 
 ## Upcoming
 
+### Added
+
+- `Setting` abstraction for config entry
+
 ### Changed
 
 - Swapped default keybinds of `Valve1Key` and `Valve3Key` to align with true finger position
 - Directly sets `t` flag in `BugleSFX` without reflection
+- Simplified config usage
 
 ## [0.3.1] - 2025-08-12